Added ekstra config. bigdecimal converted.fixed all ekstras.Added exception control

+++ src/main/java/com/work/dto/InstallmentResponseDTO.java
@@ -2,13 +2,14 @@ package com.work.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 public class InstallmentResponseDTO {
     private Long id;
-    private Double amount;
-    private Double paidAmount;
+    private BigDecimal amount;
+    private BigDecimal paidAmount;
     private LocalDate dueDate;
     private LocalDate paymentDate;
     private Boolean isPaid;
