+++ src/main/java/com/work/dto/LoanResponseDTO.java
@@ -2,11 +2,12 @@ package com.work.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 @Data
 public class LoanResponseDTO {
     private Long loanId;
-    private Double loanAmount;
+    private BigDecimal loanAmount;
     private Integer numberOfInstallment;
     private LocalDate createDate;
     private Boolean isPaid;
